Create a CollectionAddAddon component to extract the display and logic for the AutoSearchInput for adding an add-on to a collection
