+++ src/amo/components/CollectionManager/styles.scss
@@ -4,37 +4,10 @@
 @import '~amo/css/inc/vars';
 
 .CollectionManager {
-  .NoticePlaceholder {
-    margin-top: 16px;
-    min-height: 32px;
-
-    + .AutoSearchInput {
-      .AutoSearchInput-label {
-        padding-top: 4px;
-      }
-    }
-  }
-
-  .AutoSearchInput-query {
-    border: 1px solid $grey-50;
-    border-radius: $border-radius-xs;
-  }
-
   label {
     display: block;
     font-size: $font-size-s;
     padding: 24px 0 6px;
-  }
-
-  .Notice {
-    &.NoticePlaceholder-transition-appear {
-      opacity: 1;
-    }
-
-    &.NoticePlaceholder-transition-exit-active {
-      opacity: 0;
-      transition: opacity $transition-medium ease-in-out;
-    }
   }
 
   .CollectionManager-collectionName {
@@ -53,11 +26,6 @@
   textarea {
     min-height: $font-size-default * 3;
   }
-}
-
-.CollectionManager .AutoSearchInput--autocompleteIsOpen .AutoSearchInput-query {
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
 }
 
 .CollectionManager-slug {
